feature: restruct to log file and use levl to string util
